chore(media): moved call's resize logic from heading component to media component
