+++ components/tailored/core/media/Media.html
@@ -1,14 +1,14 @@
 <div id="mediastream">
   <TailoredCoreMediaHeading />
-  <div id="media">
-    <TailoredCoreMediaUser :user="{ name: 'Taurus Nix' }" talking />
-    <TailoredCoreMediaUser :user="{ name: 'Phoenix Kalindi' }" />
+  <div id="media" ref="media">
+    <TailoredCoreMediaUser v-for="item in users" :key="item.name" ref="user" :user="{ name: item.name }" :fullscren="fullscreen" />
   </div>
   <TailoredCoreMediaActionsVolume
     :volume="audio.volume"
     isPopup
     @volumeControlValueChange="volumeControlValueChange"
   />
+  <!-- {{fullscreen}} -->
   <TailoredCoreMediaActionsSettings />
   <TailoredCoreMediaActions />
 </div>
