+++ mock/users.ts
@@ -53,3 +53,68 @@ export const User = {
   last_message:
     'Apollonius of Perga laws of physics colonies paroxysm of global death Jean-François Champollion emerged into consciousness.',
 }
+
+export const CallUsers = [
+  {
+    name: 'Taurus Nix',
+    address: '0xdf9eb223bafbe5c5271415c75aecd68c21fe3d7f',
+    status: 'Some important status message',
+    state: 'online',
+    last_message:
+      'Apollonius of Perga laws of physics colonies paroxysm of global death Jean-François Champollion emerged into consciousness.',
+  },
+  {
+    name: 'Halley Themis',
+    address: '0x07ee55aa48bb72dcc6e9d78256648910de513eca',
+    status: 'Working on the space station',
+    state: 'offline',
+    last_message:
+      'At the edge of forever bits of moving fluff radio telescope the carbon in our apple pies are creatures of the cosmos rings of Uranus.',
+  },
+  // {
+  //   name: 'Lyra Cassini',
+  //   address: '0x9bf4001d307dfd62b26a2f1307ee0c0307632d59',
+  //   status: 'Working on the space station',
+  //   state: 'idle',
+  //   last_message:
+  //     'Tendrils of gossamer clouds stirred by starlight citizens of distant epochs network of wormholes corpus callosum the sky calls to us.',
+  // },
+  // {
+  //   name: 'Phoenix Kalindi',
+  //   address: '0xc61b9bb3a7a0767e3179713f3a5c7a9aedce193c',
+  //   status: 'Working on the space station',
+  //   state: 'online',
+  //   last_message:
+  //     'Ship of the imagination consciousness across the centuries network of wormholes finite but unbounded inconspicuous motes of rock and gas.',
+  //   profile_picture:
+  //     'https://st2.depositphotos.com/1009634/7235/v/600/depositphotos_72350117-stock-illustration-no-user-profile-picture-hand.jpg',
+  // },
+  // {
+  //   name: 'Ariel Larissa',
+  //   address: '0xdc76cd25977e0a5ae17155770273ad58648900d3',
+  //   status: 'Working on the space station',
+  //   state: 'idle',
+  //   last_message:
+  //     'Network of wormholes science citizens of distant epochs tingling of the spine cosmic ocean brain is the seed of intelligence. The sky calls to us two ghostly white figures in coveralls and helmets are softly dancing muse about paroxysm of global death the only home weve ever known hearts of the stars.',
+  //   profile_picture:
+  //     'https://st2.depositphotos.com/1009634/7235/v/600/depositphotos_72350117-stock-illustration-no-user-profile-picture-hand.jpg',
+  // },
+  // {
+  //   name: 'Lyra Cassini test',
+  //   address: '0x9bf4001d307dfd62b26a2f1307ee0c0307632d59',
+  //   status: 'Working on the space station',
+  //   state: 'idle',
+  //   last_message:
+  //     'Tendrils of gossamer clouds stirred by starlight citizens of distant epochs network of wormholes corpus callosum the sky calls to us.',
+  // },
+  // {
+  //   name: 'Phoenix Kalindi test',
+  //   address: '0xc61b9bb3a7a0767e3179713f3a5c7a9aedce193c',
+  //   status: 'Working on the space station',
+  //   state: 'online',
+  //   last_message:
+  //     'Ship of the imagination consciousness across the centuries network of wormholes finite but unbounded inconspicuous motes of rock and gas.',
+  //   profile_picture:
+  //     'https://st2.depositphotos.com/1009634/7235/v/600/depositphotos_72350117-stock-illustration-no-user-profile-picture-hand.jpg',
+  // },
+]
